requirements.txt
Django==2.0.7
django-cors-headers==2.4.0
django-storages==1.7.1
djangorestframework==3.8.2
gunicorn==19.8.1
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
numpy==1.14.3
Pillow==5.1.0
requests==2.22.0
scikit-learn==0.21.2
scipy==1.1.0
sklearn==0.0
tensorflow==1.13.1